Add some shortcut functions
There is an issue with the overlay that affects IE users that I've tried to solve, but I think I'm going to ignore.  It
only affects IE10 & 11 users who are navigating with a keyboard and can't use a mouse, and doesn't even break the page for
them unless they also rely on a screenreader, and it's apparently fixed in Edge.

The problem is that the overlay's Close link won't get focus if, when you open the overlay, the cursor is in an input
field that doesn't have a letter typed in it.  Apparently IE is too liberal about firing input events on fields that have
placeholders.  This issue is documented here:

https://connect.microsoft.com/IE/feedback/details/856700/ie11-ie10-send-input-events-at-times-when-it-is-not-appropriate